# src/main/java/ar/edu/utn/frc/tup/lciii/dtos/SaveCountriesDTO.java
package ar.edu.utn.frc.tup.lciii.dtos;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@NoArgsConstructor
@AllArgsConstructor
public class SaveCountriesDTO {
  private Integer amountOfCountryToSave;
}